pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "matteros"
version = "0.1.0"
description = "CLI-first legal ops orchestration with safe, auditable agent workflows"
readme = "README.md"
requires-python = ">=3.12"
license = "AGPL-3.0-only"
authors = [
  { name = "MatterOS Contributors" }
]
dependencies = [
  "typer>=0.12.3",
  "pydantic>=2.8.2",
  "PyYAML>=6.0.2",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "vcrpy>=6.0.2"
]

[project.scripts]
matteros = "matteros.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
  "external: external API contract tests (opt-in)"
]

[tool.hatch.build.targets.wheel]
packages = ["matteros"]
